Move MulticastServer to utils
